Add padding snappy message test for v2
